refactor: inputParameter 타입 전역으로 정의
